fix:move files and folders to update the project architecture
