+++ src/app/(dashboard)/trip/page.tsx
