+++ src/components/primitives/Calendar/Calendar.tsx
@@ -1,19 +1,19 @@
-import { getDate, getRangeDate } from "@/src/utils/date";
-import { cn } from "@/src/utils/twMerge";
+import { getDate, getRangeDate } from "@/src/lib/utils/date";
+import { cn } from "@/src/lib/utils/twMerge";
 import { Calendar as CalendarIcon } from "lucide-react";
 import { ButtonHTMLAttributes, useEffect, useRef, useState } from "react";
 import {
-  Calendar as ReactCalendar,
-  CalendarProps as ReactCalendarProps,
+    Calendar as ReactCalendar,
+    CalendarProps as ReactCalendarProps,
 } from "react-calendar";
 import "react-calendar/dist/Calendar.css";
 import { VariantProps } from "tailwind-variants";
 import { Button } from "../Button/Button";
 import {
-  calendarVariants,
-  sCalendar,
-  sCalendarWrapper,
-  sWrapper,
+    calendarVariants,
+    sCalendar,
+    sCalendarWrapper,
+    sWrapper,
 } from "./Calendar.variants";
 
 type ValuePiece = Date | null;
