+++ src/lib/exceptions/default-error.ts
