+++ src/lib/exceptions/http-error.ts
