+++ src/lib/utils/text.ts
